+++ core/crates/vm/husky-vm/src/entity.rs
@@ -2,14 +2,14 @@ use crate::*;
 use std::sync::atomic::{AtomicUsize, Ordering};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct EntityUid(usize);
+pub struct EntityUid(u64);
 
 impl EntityUid {
-    pub unsafe fn from_raw(raw: usize) -> EntityUid {
+    pub unsafe fn from_raw(raw: u64) -> EntityUid {
         Self(raw)
     }
 
-    pub fn raw(&self) -> usize {
+    pub fn raw(&self) -> u64 {
         self.0
     }
 }
